Fix purchasing to always have the dialog appear
Sometimes, a weird behavior in the Xcode simulator occurs where the dialog does not always appear and instead immediately pushes through the transaction without confirming with the user. The problem is that the transaction is not marked as finished explicitly.

After handling successful transactions, always end it by using `transaction.finish()`.